[DT-1293] Refactor GetUserAgent to GetHeaders
