+++ app-games/src/main/java/com/aptoide/android/aptoidegames/network/AptoideGetHeaders.kt
@@ -1,22 +1,23 @@
-package cm.aptoide.pt.network
+package com.aptoide.android.aptoidegames.network
 
 import android.content.pm.PackageManager
-import cm.aptoide.pt.BuildConfig
+import com.aptoide.android.aptoidegames.BuildConfig
+import com.aptoide.android.aptoidegames.network.repository.IdsRepository
 import cm.aptoide.pt.aptoide_network.data.network.GetUserAgent
 import cm.aptoide.pt.environment_info.DeviceInfo
 import cm.aptoide.pt.extensions.calculateMD5
 import cm.aptoide.pt.extensions.getPackageInfo
-import cm.aptoide.pt.network.repository.IdsRepository
+import com.appcoins.payments.network.RestClientInjectParams
 import java.io.File
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class AptoideGetUserAgent @Inject constructor(
+class AptoideGetHeaders @Inject constructor(
   private val packageManager: PackageManager,
   private val idsRepository: IdsRepository,
   private val deviceInfo: DeviceInfo,
-) : GetUserAgent {
+) : GetUserAgent, RestClientInjectParams {
 
   private val versionName = BuildConfig.VERSION_NAME
   private val aptoidePackage = BuildConfig.APPLICATION_ID
@@ -31,8 +32,8 @@ class AptoideGetUserAgent @Inject constructor(
       ?: "None"
   }
 
-  override operator fun invoke(): String =
-    "Aptoide/${versionName} " +
+  override fun getUserAgent(): String =
+    "AppGames/${versionName} " +
       "(Linux; Android ${deviceInfo.getAndroidRelease()}; " +
       "${deviceInfo.getApiLevel()}; " +
       "${deviceInfo.getModel()} " +
@@ -43,4 +44,8 @@ class AptoideGetUserAgent @Inject constructor(
       "${cachedMd5}; " +
       "${deviceInfo.getScreenDimensions()};" +
       "${idsRepository.aptoideClientUuid})"
+
+  override fun getUserAgent(): String = invoke()
+
+  override val channel: String = "aptoide-games"
 }
