+++ app/src/main/java/cm/aptoide/pt/network/AptoideGetHeaders.kt
@@ -1,23 +1,22 @@
-package com.aptoide.android.aptoidegames.network
+package cm.aptoide.pt.network
 
 import android.content.pm.PackageManager
-import com.aptoide.android.aptoidegames.BuildConfig
-import com.aptoide.android.aptoidegames.network.repository.IdsRepository
+import cm.aptoide.pt.BuildConfig
 import cm.aptoide.pt.aptoide_network.data.network.GetUserAgent
 import cm.aptoide.pt.environment_info.DeviceInfo
 import cm.aptoide.pt.extensions.calculateMD5
 import cm.aptoide.pt.extensions.getPackageInfo
-import com.appcoins.payments.network.RestClientInjectParams
+import cm.aptoide.pt.network.repository.IdsRepository
 import java.io.File
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class AptoideGetUserAgent @Inject constructor(
+class AptoideGetHeaders @Inject constructor(
   private val packageManager: PackageManager,
   private val idsRepository: IdsRepository,
   private val deviceInfo: DeviceInfo,
-) : GetUserAgent, RestClientInjectParams {
+) : GetUserAgent {
 
   private val versionName = BuildConfig.VERSION_NAME
   private val aptoidePackage = BuildConfig.APPLICATION_ID
@@ -32,8 +31,8 @@ class AptoideGetUserAgent @Inject constructor(
       ?: "None"
   }
 
-  override operator fun invoke(): String =
-    "AppGames/${versionName} " +
+  override fun getUserAgent(): String =
+    "Aptoide/${versionName} " +
       "(Linux; Android ${deviceInfo.getAndroidRelease()}; " +
       "${deviceInfo.getApiLevel()}; " +
       "${deviceInfo.getModel()} " +
@@ -44,8 +43,4 @@ class AptoideGetUserAgent @Inject constructor(
       "${cachedMd5}; " +
       "${deviceInfo.getScreenDimensions()};" +
       "${idsRepository.aptoideClientUuid})"
-
-  override fun getUserAgent(): String = invoke()
-
-  override val channel: String = "aptoide-games"
 }
